requirements.txt
alembic==1.13.1
asgiref==3.8.1
attrs==23.2.0
Authlib==1.3.1
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.13.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Limiter==3.7.0
Flask-Migrate==4.0.7
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.12.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==2.1.0
ordered-set==4.1.0
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.3
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==1.1.0
rich==13.7.1
shortuuid==1.0.13
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.2
validators==0.28.3
Werkzeug==3.0.3
wrapt==1.16.0
Zenora==0.0.3.post1
